Binary Search: find a target number in an array of integers
For this exercise, we could loop through the nums list and check if the element
is equal target and, if so, return its index (solution1.py).
However, this solution is O(n), and does not attend the requirement of complexity
being O(log n), present in solution2.py.

Reference: https://leetcode.com/problems/binary-search